Taken into account pull request's comments. Also, created abstract StatisticalFilter, for reduce redundant code in statistical type filters.
