+++ jis-api/src/main/java/pl/edu/uj/JImageStream/model/StreamableImage.java
@@ -8,12 +8,11 @@ import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.IOException;
 
-//todo needs some added functionality, return bufferedImage, etc
+//todo needs some added functionality, etc
 public class StreamableImage {
 
     private BufferedImage bufferedImage;
 
-    //todo extract to IOImageUtilsClass, maybe
     public StreamableImage(File file) throws IOException {
         bufferedImage = ImageUtils.convert(ImageIO.read(file), BufferedImage.TYPE_INT_ARGB);
     }
@@ -22,19 +21,14 @@ public class StreamableImage {
         this.bufferedImage = bufferedImage;
     }
 
-    //todo same, maybe
+    //todo refactor, move to another class
     public void save(String format, String file) throws IOException {
         ImageIO.write(bufferedImage, format, new File(file));
     }
 
-    public int getHeight() {
-        return bufferedImage.getHeight();
+    public BufferedImage getBufferedImage() {
+        return bufferedImage;
     }
-
-    public int getWidth() {
-        return bufferedImage.getWidth();
-    }
-
 
     public ImageStream stream() {
         return new ImageStream(bufferedImage, false);
