+++ jis-documentation/src/test/java/pl/edu/uj/JImageStream/tests/filters/BlueFilterTest.java
@@ -1,22 +1,25 @@
+package pl.edu.uj.JImageStream.tests.filters;
+
 import org.junit.Test;
 import pl.edu.uj.JImageStream.collectors.BufferedImageCollector;
-import pl.edu.uj.JImageStream.filters.color.GreenFilter;
+import pl.edu.uj.JImageStream.filters.color.BlueFilter;
 
 import javax.imageio.ImageIO;
 import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.IOException;
 
-public class GreenFilterTest extends AbstractBaseTest {
+public class BlueFilterTest extends AbstractBaseTest {
     @Test
-    public void greenFilterTest(){
-
+    public void blueFilterTest() {
+        // tag::blueFilter[]
         BufferedImage bufferedImage = streamableImage.stream()
-                .apply(new GreenFilter())
+                .apply(new BlueFilter())
                 .collect(new BufferedImageCollector());
+        // end::blueFilter[]
 
         try {
-            ImageIO.write(bufferedImage, "png", new File("target/docs/images/GreenFilter.png"));
+            ImageIO.write(bufferedImage, "png", new File("target/docs/images/BlueFilter.png"));
         } catch (IOException e) {
             e.printStackTrace();
         }
