+++ jis-documentation/src/test/java/pl/edu/uj/JImageStream/tests/filters/EmbossFilterTest.java
@@ -1,3 +1,5 @@
+package pl.edu.uj.JImageStream.tests.filters;
+
 import org.junit.Test;
 import pl.edu.uj.JImageStream.collectors.BufferedImageCollector;
 import pl.edu.uj.JImageStream.filters.convolve.EmbossFilter;
@@ -9,7 +11,7 @@ import java.io.IOException;
 
 public class EmbossFilterTest extends AbstractBaseTest {
     @Test
-    public void embossFilterTest(){
+    public void embossFilterTest() {
 
         BufferedImage bufferedImage = streamableImage.stream()
                 .apply(new EmbossFilter())
