+++ jis-documentation/src/test/java/pl/edu/uj/JImageStream/tests/filters/GaussFilterTest.java
@@ -1,3 +1,5 @@
+package pl.edu.uj.JImageStream.tests.filters;
+
 import org.junit.Test;
 import pl.edu.uj.JImageStream.collectors.BufferedImageCollector;
 import pl.edu.uj.JImageStream.filters.noise.GaussFilter;
