+++ jis-documentation/src/test/java/pl/edu/uj/JImageStream/tests/filters/MaxFilterTest.java
@@ -1,21 +1,24 @@
+package pl.edu.uj.JImageStream.tests.filters;
+
 import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.IOException;
 import javax.imageio.ImageIO;
+
 import org.junit.Test;
 import pl.edu.uj.JImageStream.collectors.BufferedImageCollector;
-import pl.edu.uj.JImageStream.filters.statistical.MedianFilter;
+import pl.edu.uj.JImageStream.filters.statistical.MaxFilter;
 
-public class MedianFilterTest extends AbstractBaseTest {
+public class MaxFilterTest extends AbstractBaseTest {
 
     @Test
-    public void medianTest() {
+    public void maxTest() {
         BufferedImage bufferedImage = streamableImage.parallelStream()
-                .setThreads(50).apply(new MedianFilter(21))
+                .setThreads(50).apply(new MaxFilter(9))
                 .collect(new BufferedImageCollector());
 
         try {
-            ImageIO.write(bufferedImage, "png", new File("target/docs/images/MedianFilter.png"));
+            ImageIO.write(bufferedImage, "png", new File("target/docs/images/MaxFilter.png"));
         } catch (IOException e) {
             e.printStackTrace();
         }
