+++ jis-documentation/src/test/java/pl/edu/uj/JImageStream/tests/filters/MinFilterTest.java
@@ -1,21 +1,24 @@
+package pl.edu.uj.JImageStream.tests.filters;
+
 import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.IOException;
 import javax.imageio.ImageIO;
+
 import org.junit.Test;
 import pl.edu.uj.JImageStream.collectors.BufferedImageCollector;
-import pl.edu.uj.JImageStream.filters.statistical.MeanFilter;
+import pl.edu.uj.JImageStream.filters.statistical.MinFilter;
 
-public class MeanFilterTest extends AbstractBaseTest {
+public class MinFilterTest extends AbstractBaseTest {
 
     @Test
-    public void meanTest() {
+    public void minTest() {
         BufferedImage bufferedImage = streamableImage.parallelStream()
-                .setThreads(100).apply(new MeanFilter(15))
+                .setThreads(50).apply(new MinFilter(9))
                 .collect(new BufferedImageCollector());
 
         try {
-            ImageIO.write(bufferedImage, "png", new File("target/docs/images/MeanFilter.png"));
+            ImageIO.write(bufferedImage, "png", new File("target/docs/images/MinFilter.png"));
         } catch (IOException e) {
             e.printStackTrace();
         }
