+++ jis-plugins/src/main/java/pl/edu/uj/JImageStream/filters/color/SepiaFilter.java
@@ -10,8 +10,7 @@ public class SepiaFilter extends Filter {
         int outputRed = (int) ((input.getRed() * .393) + (input.getGreen() * .769) + (input.getBlue() * .189));
         int outputGreen = (int) ((input.getRed() * .349) + (input.getGreen() * .686) + (input.getBlue() * .168));
         int outputBlue = (int) ((input.getRed() * .272) + (input.getGreen() * .534) + (input.getBlue() * .131));
-        int alpha = getPixel(x, y).getAlpha();
-        setPixel(x, y, new Pixel(Math.min(outputRed, 255), Math.min(outputGreen, 255), Math.min(outputBlue, 255), alpha));
+        setPixel(x, y, new Pixel(outputRed, outputGreen, outputBlue, input.getAlpha()));
 
     }
 }
