+++ jis-plugins/src/main/java/pl/edu/uj/JImageStream/filters/convolve/EmbossFilter.java
@@ -6,11 +6,11 @@ public class EmbossFilter extends ConvolutionFilter {
 
     public EmbossFilter() {
         setKernelSize(3);
-        createKernel(3);
+        createKernel();
     }
 
     @Override
-    protected void createKernel(int s) {
+    protected void createKernel() {
         this.kernel = new float[]{
                 -2.0f, 0.0f, 0.0f,
                 0.0f, 1.0f, 0.0f,
