+++ jis-plugins/src/main/java/pl/edu/uj/JImageStream/filters/edge/RobertsCrossXFilter.java
@@ -5,17 +5,17 @@ import pl.edu.uj.JImageStream.filters.ConvolutionFilter;
 public class RobertsCrossXFilter extends ConvolutionFilter {
 
     public RobertsCrossXFilter() {
-        setKernelSize(1);
-        createKernel(1);
+        setKernelSize(2);
+        createKernel();
     }
 
     @Override
-    protected void createKernel(int s) {
+    protected void createKernel() {
         kernel = new float[]{1, 0, 0, -1};
     }
 
     @Override
     protected void setKernelSize(int s) {
-        kernelSize = 2;
+        kernelSize = s;
     }
 }
