+++ jis-plugins/src/main/java/pl/edu/uj/JImageStream/filters/edge/RobertsCrossYFilter.java
@@ -5,17 +5,17 @@ import pl.edu.uj.JImageStream.filters.ConvolutionFilter;
 public class RobertsCrossYFilter extends ConvolutionFilter {
 
     public RobertsCrossYFilter() {
-        setKernelSize(1);
-        createKernel(1);
+        setKernelSize(2);
+        createKernel();
     }
 
     @Override
-    protected void createKernel(int s) {
+    protected void createKernel() {
         kernel = new float[]{0, 1, -1, 0};
     }
 
     @Override
     protected void setKernelSize(int s) {
-        kernelSize = 2;
+        kernelSize = s;
     }
 }
