+++ jis-plugins/src/main/java/pl/edu/uj/JImageStream/filters/edge/SobelYFilter.java
@@ -5,20 +5,20 @@ import pl.edu.uj.JImageStream.filters.ConvolutionFilter;
 public class SobelYFilter extends ConvolutionFilter {
 
     public SobelYFilter() {
-        setKernelSize(1);
-        createKernel(1);
+        setKernelSize(3);
+        createKernel();
     }
 
     @Override
-    protected void createKernel(int s) {
+    protected void createKernel() {
         kernel = new float[]
-                { -1, -2, -1,
-                  0, 0, 0,
-                  1, 2, 1 };
+                {-1, -2, -1,
+                        0, 0, 0,
+                        1, 2, 1};
     }
 
     @Override
     protected void setKernelSize(int s) {
-        kernelSize = 3;
+        kernelSize = s;
     }
 }
