+++ jis-plugins/src/main/java/pl/edu/uj/JImageStream/filters/statistical/MeanFilter.java
@@ -1,11 +1,11 @@
 package pl.edu.uj.JImageStream.filters.statistical;
 
-import java.util.ArrayList;
 import java.util.List;
-import pl.edu.uj.JImageStream.api.core.Filter;
+
+import pl.edu.uj.JImageStream.filters.StatisticalFilter;
 import pl.edu.uj.JImageStream.model.Pixel;
 
-public class MeanFilter extends Filter {
+public class MeanFilter extends StatisticalFilter {
 
     public MeanFilter() {
         maskSize = 3;
@@ -15,21 +15,9 @@ public class MeanFilter extends Filter {
         this.maskSize = maskSize;
     }
 
-    private int maskSize;
-
     @Override
     public void apply(int x, int y) {
-        List<Pixel> pixelList = new ArrayList<>();
-
-        for (int i = -maskSize / 2; i <= maskSize / 2; ++i) {
-            for (int j = -maskSize / 2; j <= maskSize / 2; ++j) {
-                try {
-                    pixelList.add(getPixel(x + i, y + j));
-                } catch (Exception e) {
-                    //corner case, no need to handle
-                }
-            }
-        }
+        List<Pixel> pixelList = getPixelList(x, y);
 
         double red = pixelList.stream().map(Pixel::getRed).mapToDouble(a -> a).average().getAsDouble();
         double green = pixelList.stream().map(Pixel::getGreen).mapToDouble(a -> a).average().getAsDouble();
