+++ jis-plugins/src/main/java/pl/edu/uj/JImageStream/filters/statistical/MedianFilter.java
@@ -1,12 +1,12 @@
 package pl.edu.uj.JImageStream.filters.statistical;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
-import pl.edu.uj.JImageStream.api.core.Filter;
+
+import pl.edu.uj.JImageStream.filters.StatisticalFilter;
 import pl.edu.uj.JImageStream.model.Pixel;
 
-public class MedianFilter extends Filter {
+public class MedianFilter extends StatisticalFilter {
 
     public MedianFilter() {
         maskSize = 3;
@@ -16,21 +16,9 @@ public class MedianFilter extends Filter {
         this.maskSize = maskSize;
     }
 
-    private int maskSize;
-
     @Override
     public void apply(int x, int y) {
-        List<Pixel> pixelList = new ArrayList<>();
-
-        for (int i = -maskSize / 2; i <= maskSize / 2; ++i) {
-            for (int j = -maskSize / 2; j <= maskSize / 2; ++j) {
-                try {
-                    pixelList.add(getPixel(x + i, y + j));
-                } catch (Exception e) {
-                    //corner case, no need to handle
-                }
-            }
-        }
+        List<Pixel> pixelList = getPixelList(x, y);
 
         List<Integer> red = pixelList.stream().map(Pixel::getRed).sorted().collect(Collectors.toList());
         List<Integer> green = pixelList.stream().map(Pixel::getGreen).sorted().collect(Collectors.toList());
@@ -43,6 +31,5 @@ public class MedianFilter extends Filter {
                         blue.get(blue.size() / 2),
                         alpha.get(alpha.size() / 2)));
     }
-
 
 }
